backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
numpy
opencv-python-headless
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
setuptools
pandas
openpyxl 